Dialogue System Final
Dialogue system recrated.
Create a list with char, icon and line than the script takes these as a queue and shows on the screen, player cannot move till the dialogue end. You need an npc with collider to create a dialogue, when the collider triggered dialogue starts and ends when the lines are finished.